requirements.txt
streamlit>=1.20
tensorflow>=2.10
Pillow>=9.0
numpy>=1.22
